@@ -5,13 +5,22 @@ description: >-
 homepage: https://github.com/dart-lang/json_rpc_2
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  stack_trace: ^1.0.0
+  stack_trace: ^1.10.0
   stream_channel: ">=1.1.0 <3.0.0"
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.0.0
-  web_socket_channel: ^1.1.0
+  pedantic: ^1.10.0
+  test: ^1.15.7
+  web_socket_channel:
+    git:
+      url: 'git://github.com/dart-lang/web_socket_channel.git'
+      ref: '76931ea1b81ba71e8319330c35285d3e88566315'
+
+dependency_overrides:
+  web_socket_channel:
+    git:
+      url: 'git://github.com/dart-lang/web_socket_channel.git'
+      ref: '76931ea1b81ba71e8319330c35285d3e88566315'
